Use new test syntax for FileLinks component
